+++ test/application_mock.dart
@@ -2,12 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:lightmeter/generated/l10n.dart';
 import 'package:lightmeter/res/theme.dart';
+import 'package:m3_lightmeter_iap/m3_lightmeter_iap.dart';
 
 /// Provides [MaterialApp] with default theme and "en" localization
 class WidgetTestApplicationMock extends StatelessWidget {
+  final IAPProductStatus productStatus;
   final Widget child;
 
-  const WidgetTestApplicationMock({required this.child, super.key});
+  const WidgetTestApplicationMock({
+    this.productStatus = IAPProductStatus.purchased,
+    required this.child,
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
